+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 - Fixed creative mode item hiding running even when the server has the mod.
 - Fixed Advanced Barrier blocks not respecting config settings.
 - Fixed Advanced Barrier blocks not turning visible when a barrier is held.
+- Fixed Advanced Barrier blocks not culling correctly when next to another advanced barrier block of the same type.